correct dtype

+++ pycbc/noise/reproduceable.py
@@ -72,7 +72,7 @@ def normal(start, end, sample_rate=16384, seed=0):
     # This is reproduceable because we used fixed seeds from known values
     block_dur = BLOCK_SAMPLES / sample_rate
     s = int(start / block_dur)
-    e = numpy.floor(end / block_dur)
+    e = int(numpy.floor(end / block_dur))
 
     # The data evenly divides so the last block would be superfluous
     if end % block_dur == 0:
